Moved integration tests to seperate files
